requirements.txt
numpy==1.23.5
pandas==1.5.3
rdkit-pypi
scikit-learn==1.3.2
ipython==8.12.0
streamlit==1.41.1
mordred==1.2.0
matplotlib==3.7.1
shap==0.47.1
Pillow==10.2.0
#git+https://github.com/streamlit/streamlit-ketcher.git
streamlit_ketcher-0.0.2-py3-none-any.whl
